feat: add layout hash and SVG retrieval endpoints to prevent redundant uploads
- Added API endpoints to retrieve the current layout hash and SVG representation.
- Introduced functions `_get_current_layout_hash` and `_get_current_layout_svg` to handle layout state queries.
- Implemented `layout_svg_generator.py` to convert .h5 layout files into interactive SVG maps with animations and metadata.
- Enabled automatic metadata generation for SVG files, including layout specifications and element counts.
- Prevented layout replacement and SVG regeneration when the uploaded layout matches the existing one (verified via hash comparison).
