+++ server/api_navimall/crud_path_optimization.py
@@ -25,68 +25,25 @@ logger = logging.getLogger(__name__)
 
 
 async def _upload_store_layout(layout_file: UploadFile, user_info: dict):
-    """
-    Upload store layout HDF5 file.
+    """Upload store layout HDF5 file using StoreLayoutManager logic."""
+    result = await StoreLayoutManager.upload_layout(layout_file, user_info)
+    logger.info(
+        "Layout upload processed for user=%s, hash=%s, updated=%s",
+        user_info.get("user"),
+        result.get("layout_hash"),
+        result.get("updated"),
+    )
+    return result
 
-    The HDF5 file should contain:
-    - 'layout': numpy array with 0=navigable, 1=POI, -1=obstacle
-    - 'edge_length': size of one grid cell in centimeters
-    """
-    try:
-        # Validate file format
-        if not layout_file.filename.endswith(".h5"):
-            raise HTTPException(
-                status_code=422, detail="File must be in HDF5 format (.h5)"
-            )
 
-        # Save uploaded file temporarily
-        temp_path = os.path.join("assets/layouts", f"temp_{layout_file.filename}")
+def _get_current_layout_hash(user_info: dict):
+    """Return information about the current layout hash."""
+    return StoreLayoutManager.get_current_layout_hash_info()
 
-        with open(temp_path, "wb") as buffer:
-            content = await layout_file.read()
-            buffer.write(content)
 
-        # Validate the file content
-        try:
-            layout, edge_length, _ = load_layout_from_h5(temp_path)
-
-            # Move to permanent location with hash-based name
-            cache_manager = StoreLayoutManager(layout)
-            layout_hash = cache_manager.get_current_hash()
-
-            permanent_path = os.path.join("assets/layouts", f"{layout_hash}.h5")
-            os.rename(temp_path, permanent_path)
-
-            # Save hash for tracking
-            hash_file = os.path.join("assets/cache", "current_layout.json")
-            save_hash_to_json(layout_hash, hash_file)
-
-            logger.info(
-                f"Layout uploaded successfully by user={user_info['user']}, "
-                f"hash={layout_hash}, shape={layout.shape}"
-            )
-
-            return {
-                "success": True,
-                "layout_hash": layout_hash,
-                "layout_shape": layout.shape,
-                "edge_length": edge_length,
-                "obstacles_count": int((layout == -1).sum()),
-                "navigable_cells": int((layout >= 0).sum()),
-                "filename": permanent_path,
-            }
-
-        except Exception as e:
-            # Clean up temp file if validation fails
-            if os.path.exists(temp_path):
-                os.remove(temp_path)
-            raise HTTPException(status_code=422, detail=f"Invalid HDF5 file: {str(e)}")
-
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Layout upload error: {str(e)}")
-        raise HTTPException(status_code=500, detail=f"Upload failed: {str(e)}")
+def _get_current_layout_svg(user_info: dict):
+    """Return information about the current layout SVG asset."""
+    return StoreLayoutManager.get_current_svg_info()
 
 
 def _get_layout_status(user_info: dict):
@@ -129,146 +86,135 @@ def _get_layout_status(user_info: dict):
 
 
 def _optimize_shopping_path(request: PathOptimizationRequest, user_info: dict):
-    """
-    Compute optimal shopping path for given POI coordinates.
-
-    Requires a store layout to be uploaded first via /upload_layout.
-    """
+    """Compute optimal shopping path for given POI coordinates."""
     import time
 
     start_time = time.time()
 
+    # Load hash of current layout
+    hash_file = os.path.join("assets/cache", "current_layout.json")
+    current_hash = load_hash_from_json(hash_file)
+
+    if not current_hash:
+        raise HTTPException(
+            status_code=422,
+            detail="No store layout uploaded. Please upload layout first.",
+        )
+
+    layout_path = os.path.join("assets/layouts", f"{current_hash}.h5")
+    if not os.path.exists(layout_path):
+        raise HTTPException(
+            status_code=422,
+            detail="Store layout file not found. Please re-upload layout.",
+        )
+
     try:
-        # Load current layout
-        hash_file = os.path.join("assets/cache", "current_layout.json")
-        current_hash = load_hash_from_json(hash_file)
+        layout, edge_length, _ = load_layout_from_h5(layout_path)
+    except Exception as exc:
+        logger.error("Failed to load layout: %s", exc)
+        raise HTTPException(status_code=500, detail="Invalid store layout file")
 
-        if not current_hash:
-            raise HTTPException(
-                status_code=422,
-                detail="No store layout uploaded. Please upload layout first.",
-            )
-
-        layout_file = os.path.join("assets/layouts", f"{current_hash}.h5")
-        if not os.path.exists(layout_file):
-            raise HTTPException(
-                status_code=422,
-                detail="Store layout file not found. Please re-upload layout.",
-            )
-
-        # Load layout and edge length
-        layout, edge_length, _ = load_layout_from_h5(layout_file)
-
-        # Convert POI coordinates to list of tuples
-        poi_coords_real = [(poi.x, poi.y) for poi in request.poi_coordinates]
-
-        if len(poi_coords_real) < 2:
-            raise HTTPException(
-                status_code=422,
-                detail="At least 2 POIs are required for path optimization",
-            )
-
-        # Step 1: Layout versioning management
-        layout_manager = StoreLayoutManager(
-            layout, current_hash, svg_assets_dir="assets/svg"
+    poi_coords_real = [(poi.x, poi.y) for poi in request.poi_coordinates]
+    if len(poi_coords_real) < 2:
+        raise HTTPException(
+            status_code=422,
+            detail="At least 2 POIs are required for path optimization",
         )
 
-        # Step 2: POI mapping
-        poi_mapper = POIMapper(
-            layout, request.distance_threshold, poi_coords_real, edge_length
-        )
-        grid_with_poi, distance_threshold_grid = poi_mapper.generate_grid()
-        poi_grid_coords = poi_mapper.get_poi_grid_coordinates()
+    layout_manager = StoreLayoutManager(
+        layout=layout,
+        previous_hash=current_hash,
+        svg_assets_dir="assets/svg",
+        layout_path=layout_path,
+        edge_length=edge_length,
+    )
 
-        # Step 3: Update svg assets if needed
-        layout_hash, svg_path = layout_manager.update_svg_if_needed()
-        generated_layout_svg = layout_hash == current_hash
+    poi_mapper = POIMapper(
+        layout, request.distance_threshold, poi_coords_real, edge_length
+    )
+    grid_with_poi, distance_threshold_grid = poi_mapper.generate_grid()
+    poi_grid_coords = poi_mapper.get_poi_grid_coordinates()
 
-        try:
-            solver = PathfindingSolverFactory.create_solver(
-                grid_with_poi=grid_with_poi,
-                distance_threshold_grid=distance_threshold_grid,
-                poi_coords=poi_grid_coords,
-                algorithm=request.pathfinding_algorithm,
-                diagonal_movement=request.diagonal_movement,
-            )
-            logger.info(
-                f"Using {request.pathfinding_algorithm.upper()} pathfinding algorithm"
-            )
-        except ImportError:
-            # Fallback to JPS if pathfinding library not available
-            logger.warning("pathfinding library not available")
+    layout_hash, _ = layout_manager.update_svg_if_needed(layout_path)
+    generated_layout_svg = layout_manager.last_svg_updated
 
-        except ValueError as e:
-            # Invalid algorithm, fallback to A*
-            logger.warning(
-                f"Invalid algorithm {request.pathfinding_algorithm}, falling back to A*"
-            )
-            solver = PathfindingSolverFactory.create_solver(
-                grid_with_poi=grid_with_poi,
-                distance_threshold_grid=distance_threshold_grid,
-                poi_coords=poi_grid_coords,
-                algorithm="astar",
-                diagonal_movement=request.diagonal_movement,
-            )
-
-        distance_matrix, path_matrix = solver.compute_all_paths()
-
-        # Step 5: TSP solving
-        tsp_solver = TSPSolver(distance_matrix, max_runtime=request.max_runtime)
-        visiting_order = tsp_solver.solve()
-
-        # Step 6: Final path generation with A* fallback support
-        path_generator = FinalPathGenerator(
-            path_matrix=path_matrix,
-            visiting_order=visiting_order,
-            poi_coords=poi_grid_coords,
+    try:
+        solver = PathfindingSolverFactory.create_solver(
             grid_with_poi=grid_with_poi,
-            pathfinding_algorithm=request.pathfinding_algorithm,
+            distance_threshold_grid=distance_threshold_grid,
+            poi_coords=poi_grid_coords,
+            algorithm=request.pathfinding_algorithm,
             diagonal_movement=request.diagonal_movement,
         )
-        complete_path = path_generator.generate_complete_path(
-            request.include_return_to_start
-        )
-
-        # Calculate results
-        total_distance = path_generator.calculate_total_distance()
-        computation_time = time.time() - start_time
-
-        # Get statistics
-        optimization_stats = solver.get_optimization_stats()
-        path_summary = path_generator.get_path_summary()
-
         logger.info(
-            f"Path optimization completed for user={user_info['user']}, "
-            f"POIs={len(poi_coords_real)}, distance={total_distance:.2f}, "
-            f"time={computation_time:.2f}s"
+            "Using %s pathfinding algorithm",
+            request.pathfinding_algorithm.upper(),
+        )
+    except ImportError:
+        logger.warning("pathfinding library not available, falling back to A*")
+        solver = PathfindingSolverFactory.create_solver(
+            grid_with_poi=grid_with_poi,
+            distance_threshold_grid=distance_threshold_grid,
+            poi_coords=poi_grid_coords,
+            algorithm="astar",
+            diagonal_movement=request.diagonal_movement,
+        )
+    except ValueError as exc:
+        logger.warning("%s. Falling back to A*", exc)
+        solver = PathfindingSolverFactory.create_solver(
+            grid_with_poi=grid_with_poi,
+            distance_threshold_grid=distance_threshold_grid,
+            poi_coords=poi_grid_coords,
+            algorithm="astar",
+            diagonal_movement=request.diagonal_movement,
         )
 
-        # Create response with cleaned data (no NumPy types)
-        response_data = {
-            "success": True,
-            "total_distance": float(total_distance),
-            "visiting_order": visiting_order,
-            "complete_path": complete_path,
-            "poi_count": len(poi_coords_real),
-            "computation_time": float(computation_time),
-            "generated_layout_svg": generated_layout_svg,
-            "layout_hash": layout_hash,
-            "optimization_stats": optimization_stats,
-            "path_summary": path_summary,
-        }
+    distance_matrix, path_matrix = solver.compute_all_paths()
 
-        # Clean all NumPy types from response
-        cleaned_response = clean_optimization_response(response_data)
+    tsp_solver = TSPSolver(distance_matrix, max_runtime=request.max_runtime)
+    visiting_order = tsp_solver.solve()
 
-        return PathOptimizationResponse(**cleaned_response)
+    path_generator = FinalPathGenerator(
+        path_matrix=path_matrix,
+        visiting_order=visiting_order,
+        poi_coords=poi_grid_coords,
+        grid_with_poi=grid_with_poi,
+        pathfinding_algorithm=request.pathfinding_algorithm,
+        diagonal_movement=request.diagonal_movement,
+    )
+    complete_path = path_generator.generate_complete_path(
+        request.include_return_to_start
+    )
 
-    except HTTPException:
-        raise
-    except Exception as e:
-        logger.error(f"Path optimization error: {str(e)}")
-        raise HTTPException(status_code=500, detail=f"Optimization failed: {str(e)}")
+    total_distance = path_generator.calculate_total_distance()
+    computation_time = time.time() - start_time
+
+    optimization_stats = solver.get_optimization_stats()
+    path_summary = path_generator.get_path_summary()
+
+    logger.info(
+        "Path optimization completed for user=%s, POIs=%d, distance=%.2f, time=%.2fs",
+        user_info["user"],
+        len(poi_coords_real),
+        total_distance,
+        computation_time,
+    )
+
+    response_data = {
+        "success": True,
+        "total_distance": float(total_distance),
+        "visiting_order": visiting_order,
+        "complete_path": complete_path,
+        "poi_count": len(poi_coords_real),
+        "computation_time": float(computation_time),
+        "generated_layout_svg": generated_layout_svg,
+        "layout_hash": layout_hash,
+        "optimization_stats": optimization_stats,
+        "path_summary": path_summary,
+    }
+
+    cleaned_response = clean_optimization_response(response_data)
+    return PathOptimizationResponse(**cleaned_response)
 
 
 def _validate_poi_placement(request: PathOptimizationRequest, user_info: dict):
